LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/resource-timing-bodySize.https-expected.txt


FAIL Response body size: constructed, cors, TAO pass assert_equals: expected 0 but got 3
FAIL Response body size: forward, cors, TAO pass assert_equals: expected 0 but got 70
FAIL Response body size: stream, cors, TAO pass assert_equals: expected 0 but got 70
FAIL Response body size: passthrough, cors, TAO pass assert_equals: expected 0 but got 12940
FAIL Response body size: constructed, cors, TAO fail assert_equals: expected 0 but got 3
FAIL Response body size: forward, cors, TAO fail assert_equals: expected 0 but got 70
FAIL Response body size: stream, cors, TAO fail assert_equals: expected 0 but got 70
FAIL Response body size: passthrough, cors, TAO fail assert_equals: expected 0 but got 12940
FAIL Response body size: constructed, no-cors, TAO pass assert_equals: expected 0 but got 3
FAIL Response body size: forward, no-cors, TAO pass assert_equals: expected 0 but got 70
FAIL Response body size: stream, no-cors, TAO pass assert_equals: expected 0 but got 70
PASS Response body size: passthrough, no-cors, TAO pass
FAIL Response body size: constructed, no-cors, TAO fail assert_equals: expected 0 but got 3
FAIL Response body size: forward, no-cors, TAO fail assert_equals: expected 0 but got 70
FAIL Response body size: stream, no-cors, TAO fail assert_equals: expected 0 but got 70
PASS Response body size: passthrough, no-cors, TAO fail
FAIL Response body size: constructed, same-origin, TAO pass assert_equals: expected 0 but got 3
PASS Response body size: forward, same-origin, TAO pass
FAIL Response body size: stream, same-origin, TAO pass assert_equals: expected 0 but got 70
PASS Response body size: passthrough, same-origin, TAO pass
FAIL Response body size: constructed, same-origin, TAO fail assert_equals: expected 0 but got 3
PASS Response body size: forward, same-origin, TAO fail
FAIL Response body size: stream, same-origin, TAO fail assert_equals: expected 0 but got 70
PASS Response body size: passthrough, same-origin, TAO fail
